Adds 10 and 11. Also started using linter that changed formatting of code to more C++ friendly convention.

+++ 01_loop/loop.cpp
@@ -2,8 +2,9 @@
 
 int main()
 {
-    for (int i = 0; i <= 10; i++) {
-        std::cout << "i is: " << i << std::endl;  
+    for (size_t i = 0; i <= 10; i++)
+    {
+        std::cout << "i is: " << i << std::endl;
     }
     return 0;
 }
